Make updates to TrackingAllocator, move to util
